pyproject.toml
[project]
name = "enco-test"
version = "0.1.0"
description = "Тестовое задание"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.1.3",
    "boto3>=1.42.9",
    "notebook>=7.5.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.44",
]
